fix postgres and logic

+++ internal/app/grpc/delete_last_product.go
@@ -1,15 +1,20 @@
-package merch_store
+package pvz
 
 import (
 	"context"
 	"errors"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+	"pvz/internal/auth"
 	"pvz/internal/usecase/pvz"
 	pvzpb "pvz/pkg/api/v1"
 )
 
 func (s *Service) DeleteLastProduct(ctx context.Context, request *pvzpb.DeleteLastProductRequest) (*pvzpb.DeleteLastProductResponse, error) {
+	if auth.GetUserRoleFromCtx(ctx) != employeeRole {
+		return nil, status.Error(codes.PermissionDenied, "permission denied")
+	}
+
 	err := s.PVZ.DeleteLastProduct(ctx, request.PvzId)
 	if errors.Is(err, pvz.ErrPVZNotFound) {
 		return nil, status.Error(codes.InvalidArgument, "no pvz with id "+request.PvzId)
@@ -18,7 +23,7 @@ func (s *Service) DeleteLastProduct(ctx context.Context, request *pvzpb.DeleteLa
 	} else if errors.Is(err, pvz.ErrNoProducts) {
 		return nil, status.Error(codes.FailedPrecondition, "no products in last reception in pvz with id "+request.PvzId)
 	} else if err != nil {
-		return nil, status.Error(codes.Internal, "unexpected error while deleting last product")
+		return nil, status.Errorf(codes.Internal, "unexpected error while deleting last product: %s", err.Error())
 	}
 
 	return &pvzpb.DeleteLastProductResponse{Message: "product is deleted from last opened reception"}, nil
